Add support for SSH cloning (#26)
This PR introduces SSH cloning to the CVE collector so private
repositories are supported. The script now takes a secret that contains
URLs and private keys, so when it will clone a HTTPS repository that
exists in that Secret it uses the private SSH key instead mangling the
HTTPS URL to create an SSH one (`git@<host>:<path>`).

Signed-off-by: Hector Martinez <hemartin@redhat.com>